+++ src/NumberInput.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import CustomPropTypes from './util/propTypes';
 import { number as numberLocalizer }  from './util/localizers';
 
-var format = props => numberLocalizer.getFormat('default', props.format)
+let getFormat = props => numberLocalizer.getFormat('default', props.format)
 
 export default React.createClass({
 
@@ -13,7 +13,8 @@ export default React.createClass({
     placeholder: React.PropTypes.string,
 
     format:       CustomPropTypes.numberFormat,
-    parse:        React.PropTypes.func.isRequired,
+
+    parse:        React.PropTypes.func,
     culture:      React.PropTypes.string,
 
     min:          React.PropTypes.number,
@@ -25,18 +26,23 @@ export default React.createClass({
   getDefaultProps(){
     return {
       value: null,
-      editing: false,
-      parse: (number, culture) => numberLocalizer.parse(number, culture)
+      editing: false
     }
   },
 
-  getDefaultState(props){
-    var value = props.editing
-          ? props.value
-          : formatNumber(props.value, format(props), props.culture)
+  getDefaultState(props = this.props){
+    var value = props.value
+      , decimal = numberLocalizer.decimalChar(null, props.culture)
+      , format = getFormat(props);
 
-    if ( value == null || isNaN(props.value) )
+    this._beginningWithSign = false;
+
+    if (value == null || isNaN(props.value))
       value = ''
+    else
+      value = props.editing
+        ? ('' + value).replace('.', decimal)
+        : numberLocalizer.format(value, format, props.culture)
 
     return {
       stringValue: '' + value
@@ -44,7 +50,7 @@ export default React.createClass({
   },
 
   getInitialState() {
-    return this.getDefaultState(this.props)
+    return this.getDefaultState()
   },
 
   componentWillReceiveProps(nextProps) {
@@ -61,6 +67,7 @@ export default React.createClass({
         className='rw-input'
         onChange={this._change}
         onBlur={this._finish}
+        onKeyPress={this._typing}
         aria-disabled={this.props.disabled}
         aria-readonly={this.props.readOnly}
         disabled={this.props.disabled}
@@ -70,44 +77,88 @@ export default React.createClass({
     )
   },
 
-  _change(e){
+  _typing(e) {
+    var current = e.target.value
+      , newVal = e.key;
+
+    this._beginningWithSign = current.trim() === '' && this.isSign(newVal)
+
+    this.props.onKeyPress
+      && this.props.onKeyPress(e)
+  },
+
+  _change(e) {
     var val = e.target.value
-      , number = this.props.parse(e.target.value, this.props.culture)
-      , valid = this.isValid(number);
+      , number = this._parse(e.target.value)
+      , atSign = this.isSign(val.trim())
+      , startingWithSign = this._beginningWithSign;
 
-    if( val == null || val.trim() === '' || val.trim() === '-')
+    this._beginningWithSign = false;
+
+    if (val == null || val.trim() === '' || (atSign && !startingWithSign)) {
+      this.current('')
       return this.props.onChange(null)
+    }
 
-    if( valid && number !== this.props.value && !this.isAtDelimiter(number, val))
-      return this.props.onChange(number)
+    if (this.isFlushable(number, val)) {
+      if (number !== this.props.value)
+        return this.props.onChange(number)
+      else
+        this.setState(this.getDefaultState()) // 5. -> 5
+    }
 
-    //console.log(val !== 0 && !val)
-    if ( !isNaN(number) || this.isAtDelimiter(number, val))
+    if (number < this.props.min || (atSign && startingWithSign) || this.isAtDelimiter(number, val))
       this.current(e.target.value)
   },
 
   _finish() {
     var str = this.state.stringValue
-      , number = this.props.parse(str, this.props.culture);
+      , number = this._parse(str);
 
     // if number is below the min
     // we need to flush low values and decimal stops, onBlur means i'm done inputing
-    if(!isNaN(number) && (number < this.props.min || this.isAtDelimiter(number, str)) ) {
+    if (!isNaN(number) && (number < this.props.min || this.isAtDelimiter(number, str)) ) {
       this.props.onChange(number)
     }
   },
 
-  isAtDelimiter(num, str){
-    var next;
+  _parse(strVal) {
+    let culture = this.props.culture
+      , delimChar = numberLocalizer.decimalChar(null, culture)
+      , userParse = this.props.parse;
 
-    if ( str.length <= 1) return false
+    if (userParse)
+      return userParse(strVal, culture)
 
-    next = this.props.parse(
-      str.substr(0, str.length - 1), this.props.culture)
+    strVal = strVal.replace(delimChar, '.')
+    strVal = parseFloat(strVal);
 
-    return typeof next === 'number'
-        && !isNaN(next)
-        && next === num
+    return strVal
+  },
+
+  isFlushable(num, str) {
+    return (
+          this.isValid(num)
+      && !this.isAtDelimiter(num, str)
+      && !this.isSign(str)
+    )
+  },
+
+  isSign(val) {
+    return (val || '').trim() === '-';
+  },
+
+  isAtDelimiter(num, str, props = this.props) {
+    var localeChar = numberLocalizer.decimalChar(null, props.culture)
+      , lastIndex = str.length - 1
+      , char;
+
+    if (str.length <= 1) return false
+
+    char = str[lastIndex]
+
+    return char === localeChar
+      && str.indexOf(char) === lastIndex
   },
 
   isValid(num) {
@@ -117,20 +168,8 @@ export default React.createClass({
   },
 
   //this intermediate state is for when one runs into the decimal or are typing the number
-  current(val){
+  current(val) {
     this.setState({ stringValue: val })
   }
 
 });
-
-
-// function parseLocaleFloat(number, parser, culture) {
-//   if ( typeof format === 'function')
-//     return format(number, culture)
-
-//   return config.globalize.parseFloat(number, 10, culture)
-// }
-
-function formatNumber(number, format, culture){
-  return numberLocalizer.format(number, format, culture)
-}