+++ src/util/localizers.js
@@ -42,7 +42,7 @@ function checkFormats(requiredFormats, formats){
 
 let _numberLocalizer = createWrapper('NumberPicker')
 
-export function setNumber({ format, parse, precision = () => null, formats, propType }) {
+export function setNumber({ format, parse, decimalChar = () => '.', precision = () => null, formats, propType }) {
   invariant(typeof format === 'function'
     , 'number localizer `format(..)` must be a function')
   invariant(typeof parse === 'function'
@@ -50,17 +50,20 @@ export function setNumber({ format, parse, precision = () => null, formats, prop
 
   checkFormats(REQUIRED_NUMBER_FORMATS, formats)
 
+  formats.editFormat = formats.editFormat || (str => parseFloat(str));
+
   _numberLocalizer = {
     formats,
     precision,
+    decimalChar,
     propType: propType || localePropType,
 
     format(value, str, culture){
       return _format(this, format, value, str, culture)
     },
 
-    parse(value, culture) {
-      let result = parse.call(this, value, culture)
+    parse(value, culture, format) {
+      let result = parse.call(this, value, culture, format)
       invariant(result == null || typeof result === 'number'
         , 'number localizer `parse(..)` must return a number, null, or undefined')
       return result
@@ -106,6 +109,9 @@ export let number = {
   },
   format(...args){
     return _numberLocalizer.format(...args)
+  },
+  decimalChar(...args){
+    return _numberLocalizer.decimalChar(...args)
   },
   precision(...args){
     return _numberLocalizer.precision(...args)