+++ test/index.js
@@ -1,4 +1,5 @@
 'use strict';
+require('es5-shim')
 var React = require('react');
 var widgetHelpers = require('../src/util/widgetHelpers')
 var globalize = require('globalize')
@@ -7,6 +8,19 @@ require('../src/localizers/globalize')(globalize)
 
 //disable this particular optimization
 sinon.stub(widgetHelpers, 'isFirstFocusedRender', ()=> true)
+
+beforeEach(() => {
+  sinon.stub(console, 'error');
+});
+
+afterEach(function () {
+  if (typeof console.error.restore === 'function') {
+    if (console.error.called)
+      throw new Error(`${console.error.getCall(0).args[0]} \nIn '${this.currentTest.fullTitle()}'`)
+
+    console.error.restore();
+  }
+});
 
 var testsContext = require.context('../test', true, /\.browser\.(js$|jsx$)/);